pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agentcds"
version = "0.1.0"
description = "Agentic Clinical Decision Support — Academic Research Project"
requires-python = ">=3.10"

dependencies = [
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "huggingface-hub>=0.20",
    "transformers>=4.38",
    "sentence-transformers>=2.5",
    "torch>=2.1",
    "chromadb>=0.4.22",
    "httpx>=0.26",
    "fastmcp>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentcds*"]
